% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical.paintings.srilanka.R
\docType{data}
\name{historical.paintings.locations}
\alias{historical.paintings.locations}
\title{Historical Painting Locations in Sri Lanka}
\format{
A tibble with 20 rows and 0 variables:
\describe{
\item{Kingdom}{Kingdom that the painting belongs to}
\item{Location}{Name of the location}
\item{Period}{Period}
\item{Century}{Century}
\item{UNESCO_heritage}{Whether it is a UNESCO world heritage site or not}
\item{District}{District}
\item{Province}{Province}
\item{Source}{Source of information}
\item{Accessed_data}{Information accessed date}
}
}
\source{
Accessed from varion online sources. Sources are given in the data frame
}
\usage{
historical.paintings.locations
}
\description{
Information related to the historical painting locations in Sri Lanka
}
\examples{
head(historical.paintings.locations)
}
\keyword{datasets}
