% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sf_data.R
\name{load_sf_data}
\alias{load_sf_data}
\title{Load Simple Feature Objects to Plot Maps of Sri Lanka}
\usage{
load_sf_data(file.name)
}
\arguments{
\item{file.name}{The name of the \code{.rda} file to load (e.g., \code{"lka_adm1.rda"}).}
}
\value{
An \code{sf} object containing the loaded spatial data.
}
\description{
This function loads spatial \code{.rda} files stored in \verb{inst/extdata/}.
}
\examples{
lka_adm1 <- load_sf_data("lka_adm1.rda")
plot(lka_adm1)
}
