% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SriLanka-package.R
\docType{package}
\name{SriLanka-package}
\alias{SriLanka-package}
\title{SriLanka: Simple Feature Objects to Plot Sri Lanka Maps}
\description{
Provides simple features for creating map of the country, map of provinces, map of districts and map of divisional secretariats.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thiyangt/SriLanka}
  \item Report bugs at \url{https://github.com/thiyangt/SriLanka/issues}
}

}
\author{
\strong{Maintainer}: Thiyanga S. Talagala \email{ttalagala@sjp.ac.lk} (\href{https://orcid.org/0000-0002-0656-9789}{ORCID})

}
\keyword{internal}
