% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\docType{data}
\name{SriLanka}
\alias{SriLanka}
\alias{lka_adm0}
\alias{lka_adm1}
\alias{lka_adm2}
\alias{lka_adm3}
\title{Simple feature objects to plot Sri Lanka maps}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 6 columns.

An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 9 rows and 6 columns.

An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 25 rows and 6 columns.

An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 331 rows and 5 columns.
}
\usage{
lka_adm0

lka_adm1

lka_adm2

lka_adm3
}
\description{
There are four simple feature objects in this package that can be used to
plot Sri Lanka maps:
\itemize{
\item \code{lka_adm0}: Country-level boundaries
\item \code{lka_adm1}: Province-level boundaries
\item \code{lka_adm2}: District-level boundaries
\item \code{lka_adm3}: Divisional secretariats boundaries
}
}
\examples{
library(ggplot2)

# Country-level boundaries
ggplot(lka_adm0) +
  geom_sf()

# Province-level boundaries
ggplot(lka_adm1) +
  geom_sf()

# District-level boundaries
ggplot(lka_adm2) +
  geom_sf()
  
# Divisional secretariats boundaries
ggplot(lka_adm3) +
  geom_sf()
}
\references{
https://www.geoboundaries.org/data/1_3_0/shp/LKA/ (accessed: 4 March 2025)
}
\keyword{datasets}
